Fixed integer overflow when calculating maximum metadata size. This fixes #2089

+++ Duplicati/Library/Main/Operation/BackupHandler.cs
@@ -685,7 +685,7 @@ namespace Duplicati.Library.Main.Operation
                     throw new Exception("The database was attempted repaired, but the repair did not complete. This database may be incomplete and the backup process cannot continue. You may delete the local database and attempt to repair it again.");
 
                 m_blocksize = m_options.Blocksize;
-                m_maxmetadatasize = (m_blocksize / m_options.BlockhashSize) * m_blocksize;
+                m_maxmetadatasize = (m_blocksize / (long)m_options.BlockhashSize) * m_blocksize;
 
                 m_blockbuffer = new byte[m_options.Blocksize * Math.Max(1, m_options.FileReadBufferSize / m_options.Blocksize)];
                 m_blocklistbuffer = new byte[m_options.Blocksize];
